refactor(structure): move root component into app folder and define future routes
